+++ doc/release-notes.md
@@ -3,3 +3,4 @@ Bitcoin ABC version 0.19.8 is now available from:
   <https://download.bitcoinabc.org/0.19.8/>
 
 This release includes the following features and fixes:
+ - Remove `getinfo` RPC in favor of `getblockchaininfo`, `getnetworkinfo` and `getwalletinfo`.